Sign from client package
